remove gyms listing page
